Exercício

+++ src/Lista1/Ex20.java
@@ -1,7 +1,3 @@
-/*
-
- */
-
 package Lista1;
 
 import java.util.Scanner;
@@ -10,7 +6,7 @@ public class Ex20 {
     public static void main(String[] args) {
         Scanner teclado = new Scanner(System.in);
         System.out.println("Digite o ano para ser verificado: ");
-        double ano = teclado.nextDouble();
+        double ano = teclado.nextInt();
 
         if (ano % 400 == 0) {
             System.out.println("Ano bissexto. ");
@@ -25,7 +21,6 @@ public class Ex20 {
                 System.out.println("nao e bissexto. ");
             }
         }
-
     }
 }
 /* jeito 2
